move gpsToUserGrid function
- Done per suggestion in PR review.
