+++ src/client/app/containers/MapChartContainer.ts
@@ -8,8 +8,7 @@ import PlotlyChart, { IPlotlyChartProps } from 'react-plotlyjs-ts';
 import { State } from '../types/redux/state';
 import {
 	calculateScaleFromEndpoints, meterDisplayableOnMap, Dimensions,
-	CartesianPoint, normalizeImageDimensions, trueNorthAngle, rotateShift,
-	GPSPoint, MapScale, meterMapInfoOk, trueNorthOrigin
+	CartesianPoint, normalizeImageDimensions, meterMapInfoOk, gpsToUserGrid
 } from '../utils/calibration';
 import * as _ from 'lodash';
 import getGraphColor from '../utils/getGraphColor';
@@ -221,30 +220,6 @@ function mapStateToProps(state: State) {
 	};
 	props.config.locale = state.admin.defaultLanguage;
 	return props;
-}
-
-/**
- * Convert the gps value to the equivalent Plotly grid coordinates on user map.
- * @param size The normalized size of the map
- * @param gps The GPS coordinate to convert
- * @param originGPS The GPS value of the origin on the true north map.
- * @param scaleOfMap The GPS degree per unit x, y on the true north map.
- * @returns x, y value of the gps point on the user map.
- */
-export function gpsToUserGrid(size: Dimensions, gps: GPSPoint, originGPS: GPSPoint, scaleOfMap: MapScale): CartesianPoint {
-	// We need the origin x, y value by starting from 0, 0 on the user map and
-	// shift/rotate into true north.
-	const originTrueNorth = trueNorthOrigin(size);
-	// Now, convert from GPS to true north grid (x, y).
-	// Calculate how far the point is from origin and then the units for this distance.
-	const gridTrueNorth: CartesianPoint = {
-		x: originTrueNorth.x + (gps.longitude - originGPS.longitude) / scaleOfMap.degreePerUnitX,
-		y: originTrueNorth.y + (gps.latitude - originGPS.latitude) / scaleOfMap.degreePerUnitY
-	};
-	// Now rotate to user map and shift so origin at bottom, left. Do - angle since going from
-	// true north to user map.
-	const userGrid: CartesianPoint = rotateShift(size, gridTrueNorth, 1, -trueNorthAngle.angle);
-	return userGrid;
 }
 
 export default connect(mapStateToProps)(PlotlyChart);
